+++ packages/widgets/user-management-widget/src/lib/widget/state/types.ts
@@ -1,6 +1,12 @@
 import { ActionReducerMapBuilder } from '@reduxjs/toolkit';
 import { Sdk } from '../api/sdk';
-import { CustomAttr, Role, SortParams, User } from '../api/types';
+import {
+  AssociatedTenant,
+  CustomAttr,
+  Role,
+  SortParams,
+  User,
+} from '../api/types';
 
 export type State = {
   usersList: {
@@ -45,6 +51,11 @@ export type State = {
     loading: boolean;
     error: unknown;
     data: Role[];
+  };
+  subTenantRoles: {
+    loading: boolean;
+    error: unknown;
+    data: AssociatedTenant[];
   };
   searchParams: { text: string; sort: SortParams[] };
   selectedUsersLoginIds: string[][];