+++ packages/widgets/user-profile-widget/src/lib/widget/mixins/initMixin/initMixin.ts
@@ -12,6 +12,7 @@ import { initTotpUserAuthMethodMixin } from './initComponentsMixins/initTotpUser
 import { initUserBuiltinAttributesMixin } from './initComponentsMixins/initUserBuiltinAttributesMixin';
 import { initUserCustomAttributesMixin } from './initComponentsMixins/initUserCustomAttributesMixin';
 import { initTrustedDevicesMixin } from './initComponentsMixins/initTrustedDevicesMixin';
+import { initUserPasskeysMixin } from './initComponentsMixins/initUserPasskeysMixin';
 import { initNotificationsMixin } from './initComponentsMixins/initNotificationsMixin';
 import { initTenantSelectorMixin } from './initComponentsMixins/initTenantSelectorMixin';
 import { initGenericFlowButtonMixin } from './initComponentsMixins/initGenericFlowButtonMixin';
@@ -34,6 +35,7 @@ export const initMixin = createSingletonMixin(
       initUserBuiltinAttributesMixin,
       initLogoutMixin,
       initTrustedDevicesMixin,
+      initUserPasskeysMixin,
       initNotificationsMixin,
       initTenantSelectorMixin,
       initGenericFlowButtonMixin,